+++ src/api/index.ts
@@ -1,0 +1,6 @@
+export * from "./users";
+export * from "./spaces";
+export * from "./posts";
+export * from "./comments";
+
+export { default as api } from "./client";
