+++ src/components/Search/Search.css
@@ -164,3 +164,62 @@
   font-size: 14px;
   font-style: italic;
 }
+
+.search-section-header {
+  padding: 8px 16px;
+  background: #f0f0f0;
+  border-bottom: 1px solid #e0e0e0;
+  font-size: 13px;
+  color: #555;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+}
+
+.search-user-item {
+  padding: 12px 16px;
+}
+
+.search-user-info {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.search-user-avatar {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  overflow: hidden;
+  flex-shrink: 0;
+}
+
+.search-user-avatar img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.search-user-avatar-placeholder {
+  width: 100%;
+  height: 100%;
+  background: #9F283B;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: 600;
+  font-size: 16px;
+}
+
+.search-user-details {
+  flex: 1;
+  min-width: 0;
+}
+
+.search-user-email {
+  margin: 2px 0 0 0;
+  font-size: 13px;
+  color: #888;
+  font-style: italic;
+}
