+++ src/components/SpaceUsersModal/SpaceUsersModal.css
@@ -83,6 +83,15 @@
   color: #666;
 }
 
+.loading-state {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 40px 24px;
+  color: #666;
+}
+
 .loading-spinner {
   width: 32px;
   height: 32px;
