+++ src/pages/Post/Post.tsx
@@ -3,9 +3,8 @@ import Topbar from "../../components/Topbar/Topbar";
 import Breadcrumb from "../../components/Breadcrumb/Breadcrumb";
 import { useParams, useNavigate } from 'react-router-dom';
 import { useState, useEffect } from "react";
-import type { User } from "../../types/user";
 import { useAppContext } from "../../context/AppContext";
-import { getCurrentUser, getPostById, addCommentToPost } from "../../services/api";
+import { getPostById, addCommentToPost } from "../../api";
 import type { Post as PostType } from "../../types/post";
 import { formatPostDetailDate, formatPostDetailTime } from "../../utils/dateUtils";
 import "./Post.css";
